requirements.txt
discord.py
discord.py[voice]

